Move model specific bits into model.go
Lovely how easy that was :)
