+++ database/database.go
@@ -2,46 +2,14 @@ package database
 
 import (
 	"database/sql"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"time"
 
 	log "github.com/Sirupsen/logrus"
-	"github.com/twinj/uuid"
 
 	// Imported only for it's side effect
 	_ "github.com/mattn/go-sqlite3"
 )
-
-// Note represents a single note stored by the application
-type Note struct {
-	Content   string `json:"content"`
-	Created   string `json:"created"`
-	Encrypted bool   `json:"encrypted"`
-	Password  string `json:"password"`
-	Subject   string `json:"subject"`
-	Tags      string `json:"tags"`
-	UID       string `json:"uid"`
-	Updated   string `json:"updated"`
-}
-
-// Notes is a collection of Note objects
-type Notes []Note
-
-// ToJSON converts a (filtered) note into json fields filtered:
-// - Content: For performance reasons
-// - Password: For security reasons
-func (note Note) ToJSON() (string, error) {
-	note.Content = ""
-	note.Password = ""
-	noteJSON, err := json.MarshalIndent(note, "", "  ")
-	if err != nil {
-		log.Error("Failed to parse note into json", err)
-		return "", err
-	}
-	return string(noteJSON), err
-}
 
 // connection to a sqlite database (currently hard coded for testing)
 func connection() (*sql.DB, error) {
@@ -50,23 +18,6 @@ func connection() (*sql.DB, error) {
 		panic(err)
 	}
 	return db, err
-}
-
-// persistable prepares a note for being persisted to storage
-func persistable(note Note) (Note, error) {
-	note.Updated = Now()
-	// Generate a uuid if necessary
-	if note.UID == "" {
-		note.UID = uuid.NewV4().String()
-	}
-	// Make sure the contents are encrypted if a password is set
-	if note.Password != "" {
-		note.Content = Encrypt(note.Content, note.Password)
-		note.Encrypted = true
-	} else {
-		note.Encrypted = false
-	}
-	return note, nil
 }
 
 // DeleteByUID removes a note from storage
@@ -93,7 +44,7 @@ func DeleteByUID(uid string) error {
 
 // Create a note
 func Create(note Note) (Note, error) {
-	note, err := persistable(note)
+	note, err := Persistable(note)
 	if err != nil {
 		panic(err)
 	}
@@ -148,12 +99,6 @@ func GetContentByUID(uid string, password string) (string, error) {
 	return "", fmt.Errorf("No note found")
 }
 
-// Now is the current timestamp in time.RFC3339 format
-func Now() string {
-	t := time.Now()
-	return t.Format(time.RFC3339)
-}
-
 // Search all notes as filtered by the provided query
 func Search(query string) Notes {
 	db, _ := connection()
@@ -184,7 +129,7 @@ func Search(query string) Notes {
 
 // Update a note
 func Update(note Note) (Note, error) {
-	note, err := persistable(note)
+	note, err := Persistable(note)
 	if err != nil {
 		panic(err)
 	}
